feat: pass relative paths instead of path.join
As shown in the AVA tests, this allows passing relative-filepath strings
to `createAndDeploy` rather than requiring people to import `path`.
These paths are assumed to be relative to the package root.

These paths will be resolved using the entire array of Node's lookup
paths (`module.paths`), and if all of that fails, it will try to use the
current working directory. If it still fails, it throws a helpful error.