feat: ensure tests pass in bootstrapped project
We have to ignore the tests in the `bootstrap-starter` folder in the
normal test run, since we run `yarn test` from the monorepo root and the
relative path to `compiled-contracts/status_message.wasm` is from the
perspective of the nested `bootstrap-starter` project. So this
re-ignores this test normally.

But it also adds a new test for the freshly-created project ensuring
that the tests run properly.