docs: prefer `test` rather than `t`
AVA's docs tell people to use `t`, partially because the top-level
variable is usually named `test`. But in our context we have no such
top-level variable, and we have a lot teach to people who are so new to
all of it, so it makes more sense to use a longer variable name.